code/requirements.txt
attrs==21.4.0
bcrypt==4.0.1
blessed==1.19.1
boto3==1.26.24
botocore==1.29.24
cffi==1.15.1
chalice==1.27.3
click==8.1.3
cryptography==38.0.4
dnspython==2.2.1
inquirer==2.10.1
jmespath==1.0.1
marshmallow==3.19.0
packaging==22.0
pycparser==2.21
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.6
python-editor==1.0.4
PyYAML==6.0
readchar==4.0.3
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
